refactor(backend): access log allowlist, random MC seed, handler split
- Replace raw query logging with sanitizeQuery allowlist (closes #646)
  - Non-safe params (address etc.) are redacted to [REDACTED] in middleware
  - Remove redundant RawQuery override from GetGeocode handler
- Switch MonteCarloSimulate from fixed seed (42) to time-based seed (closes #643)
- Move HandleAreaDiscovery to area_handler.go (closes #644)
- Move HandleRenovationAnalyze + validateRenovationInput to renovation_handler.go
